requirements.txt
omegaconf==2.0.2
requests==2.24.0
discord.py[voice]==1.5.0
pandas>=1.1.4
